refactor: update component imports and adjust navigation links to use new paths
