--- client/src/app/components/button/Practise2.tsx
@@ -1,36 +1,0 @@
-// type Props = {
-//     name: string
-//     id: number
-//     age?:number
-// }
-
-// const Practice2 = ({ name, id, age }: Props) => {
-//   return (
-//     <div className="flex justify-center w-full"> {/* 💡 Center container */}
-//       <section className="relative py-16 overflow-hidden max-w-4xl w-full">
-//         {/* <!-- Title --> */}
-//         <h2 className="text-xl font-semibold px-4 mb-4 text-center">Resources</h2>
-
-//         {/* <!-- Scrollable card container --> */}
-//         <div className="relative">
-//           {/* <!-- Left/Right buttons --> */}
-//           <button className="absolute left-4 top-1/2 transform -translate-y-1/2 bg-white rounded-full shadow-md p-2">
-//             ←
-//           </button>
-//           <button className="absolute right-4 top-1/2 transform -translate-y-1/2 bg-white rounded-full shadow-md p-2">
-//             →
-//           </button>
-
-//           {/* <!-- Scrollable area --> */}
-//           <div className="flex space-x-6 overflow-x-auto px-8 scrollbar-hide">
-//             <div className="bg-white rounded-lg shadow-md w-60 h-72 flex-shrink-0"></div>
-//             <div className="bg-white rounded-lg shadow-md w-60 h-72 flex-shrink-0"></div>
-//             <div className="bg-white rounded-lg shadow-md w-60 h-72 flex-shrink-0"></div>
-//           </div>
-//         </div>
-//       </section>
-//     </div>
-//   );
-// };
-
-// export default Practice2
