+++ client/src/app/contact/page.tsx
@@ -1,4 +1,4 @@
-import ContactForm from '../components/contact-form/ContactForm'
+import ContactForm from '@/components/contact-form/ContactForm'
 
 const Contact = () => {
   return (
