+++ client/src/components/navSearchCard/navSearchCard.tsx
