+++ client/src/components/shapes/Parabola.tsx
